support_service/requirements.txt
grpcio
grpcio-tools
protobuf
openai